pyproject.toml
[tool.poetry]
name = "grpc-weather-service"
version = "0.1.0"
description = ""
authors = ["connorjcantrell <connorjcantrell@gmail.com>"]
readme = "README.md"
packages = [{include = "grpc_weather_service"}]

[tool.poetry.dependencies]
python = "^3.11"
grpcio-tools = "^1.58.0"
requests = "^2.31.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
